+++ Reinvest/Investments/src/Infrastructure/Providers/AdaptersProviders.ts
@@ -1,0 +1,21 @@
+import { ContainerInterface } from 'Container/Container';
+import { QueueSender } from 'shared/hkek-sqs/QueueSender';
+import { SimpleEventBus } from 'SimpleAggregator/EventBus/EventBus';
+import { SendToQueueEventHandler } from 'SimpleAggregator/EventBus/SendToQueueEventHandler';
+
+import { Investments } from '../..';
+
+export default class AdaptersProviders {
+  private config: Investments.Config;
+
+  constructor(config: Investments.Config) {
+    this.config = config;
+  }
+
+  public boot(container: ContainerInterface) {
+    container
+      .addAsValue(SimpleEventBus.getClassName(), new SimpleEventBus(container))
+      .addObjectFactory(QueueSender, () => new QueueSender(this.config.queue), [])
+      .addObjectFactory(SendToQueueEventHandler, (queueSender: QueueSender) => new SendToQueueEventHandler(queueSender), [QueueSender]);
+  }
+}