+++ Reinvest/LegalEntities/src/Providers/PortsProvider.ts
@@ -20,7 +20,9 @@ import { CompleteProfile } from 'LegalEntities/UseCases/CompleteProfile';
 import { CreateDraftAccount } from 'LegalEntities/UseCases/CreateDraftAccount';
 import { RemoveDraftAccount } from 'LegalEntities/UseCases/RemoveDraftAccount';
 import { TransformDraftAccountIntoRegularAccount } from 'LegalEntities/UseCases/TransformDraftAccountIntoRegularAccount';
+import { UpdateCompanyForVerification } from 'LegalEntities/UseCases/UpdateCompanyForVerification';
 import { UpdateProfileForVerification } from 'LegalEntities/UseCases/UpdateProfileForVerification';
+import { UpdateStakeholderForVerification } from 'LegalEntities/UseCases/UpdateStakeholderForVerification';
 
 export class PortsProvider {
   private config: LegalEntities.Config;
@@ -45,6 +47,6 @@ export class PortsProvider {
         TransformDraftAccountIntoRegularAccount,
         RemoveDraftAccount,
       ])
-      .addSingleton(UpdateForVerificationController, [UpdateProfileForVerification]);
+      .addSingleton(UpdateForVerificationController, [UpdateProfileForVerification, UpdateCompanyForVerification, UpdateStakeholderForVerification]);
   }
 }