+++ Reinvest/Registration/src/Port/Api/SynchronizationController.ts
@@ -54,4 +54,24 @@ export class SynchronizationController {
         return false;
     }
   }
+  public async synchronizeProfile(profileId: string): Promise<boolean> {
+    const record = await this.mappingRegistryRepository.findByProfile(profileId);
+    console.log({ recordToSync: record });
+
+    return await this.synchronizeProfileUseCase.execute(record);
+  }
+
+  public async synchronizeCompany(profileId: string, accountId: string): Promise<boolean> {
+    const record = await this.mappingRegistryRepository.getCompanyById(profileId, accountId);
+    console.log({ recordToSync: record });
+
+    return await this.synchronizeCompanyUseCase.execute(record);
+  }
+
+  public async synchronizeStakeholder(profileId: string, accountId: string, stakeholderId: string): Promise<boolean> {
+    const record = await this.mappingRegistryRepository.getStakeholderById(profileId, accountId, stakeholderId);
+    console.log({ recordToSync: record });
+
+    return await this.synchronizeCompanyUseCase.execute(record);
+  }
 }